fix: client_focus_on_cancel - it makes this work. remember prior active window, before skippy was ever activated, when cancelling with Esc key
Previously there was an oversight or bug in the code, whereby this variable wasn't getting set or used correctly.
By moving assignment of variable to place slightly before in the code, skippy can now remember what it was!
:Yum: (easy fix)

Maybe for some reason(s) this was already working after moving about with the mouse. In different areas of code.
This fix is for situation(s) where mouse isn't moved / triggering those alternate code paths. So maybe other
places in the code is already OK. (And no need to bother elsewhere)